Fix divide by 0 error when ints

+++ emccd_detect/rand_em_gain.py
@@ -72,7 +72,7 @@ def _rand_pdf(n_in, em_gain, x_max, size):
     else:
         # For n > 2 use CDF approximation
         # Use x values ranging from 0 to maximum allowable x output
-        x_axis = np.arange(0, x_max)
+        x_axis = np.arange(0, x_max).astype(float)
         x_axis[0] = np.finfo(float).eps  # Use epsilon to avoid divide by 0
         cdf = _get_cdf(n_in, em_gain, x_axis)
